[PORT FROM MAIN] Phone: Play an out of service tone if SIM card is removed during call
BZ: 11908

In the audio framework, if no tone is played at call disonnection, the output is not
released correctly in accordance to the platform audio drivers. Thus the audio is not
reset gracefully and the Modem ALSA driver not closed. As a result it is not possible
to playback a stream after a call ended due to a SIM removal. Playing an out of service
tone solves the problem.

Category: AOSP Improvement
Domain: Telephony-Phone
Origin: Internal
Upstream-Candidate: Yes

Change-Id: Ib2076a6cea8470a59887614a68c806a3702c7a09
Orig-Change-Id: Iff7ac1727588ddd0061703af468f0dd751992a2f
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/125039
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>