[PORT FROM MAIN] Phone: Play an out of service tone if SIM card is removed during call
BZ: 11908

In the audio framework, if no tone is played at call disonnection, the outpu
released correctly in accordance to the platform audio drivers. Thus the aud
reset gracefully and the Modem ALSA driver not closed. As a result it is not
to playback a stream after a call ended due to a SIM removal. Playing an out
tone solves the problem.

Change-Id: Iff7ac1727588ddd0061703af468f0dd751992a2f
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/75076
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>