[PORT FROM MAIN] Phone: SIM Contacts are not refreshed after SIM Hot Swap
BZ: 61192

This patch introduces a broadcast receiver for SIM Status and
when it is ABSENT resets the internal data, so as to re-read
the adn when activity is reloaded.

The new contacts will be loaded when the user re-enters the view.

Category: aosp improvement
Domain: AOSP.Origin-PhoneApp
Severity: medium
Origin: internal
Upstream-Candidate: yes

Change-Id: I5d0ddbb676c5ccfb928252753ba9ecc742c45bfd
Orig-Change-Id: I3efa9cf403f4347f6a73fc2a42a86ffbd1c9f129
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/123378
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>