[PORT FROM MAIN] Telephony: Call Forwarding items are grayed when wrong numbers are used
BZ: 52109

In call forwarding settings (always, busy, unanswered, unreachable), when us
enter an empty or too long phone number, a pop-up with an error message
appears and button is grayed.
If user wants to modify the settings again, it is necessary to go in sleep
mode or go back and re-enter in the menu.
With this patch, pop-up is still available and button is never grayed.
When user enter a wrong phone number, button is still available.

Change-Id: I0f92eaedc5a56cd941223ebde08dfb7e52ceaa9e
Signed-off-by: Goubert, Tony <tonyx.goubert@intel.com>
Reviewed-on: http://android.intel.com:8080/75142
Reviewed-by: Alagnou, GregoryX <gregoryx.alagnou@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: Le Meur, ArnaudX <arnaudx.le.meur@intel.com>
Reviewed-by: Blagny, ChristianX <christianx.blagny@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>