[PORT FROM MAIN] Phone: Disable SIM UIs in Airplane mode/SIM removal
BZ: 59396

If the airplane mode is activated using the power
button and if the current activity is one of the
SIM related activity, then the activity
should be disabled. Currently, it is not disabled.

Since the framework doesn't completely support
SIM access in airplane mode, there are few
issues seen. Same applies to if the SIM is removed
when the current activity is one of the SIM
related activity.

This patch disables/greys out the SIM related UIs
when the airplane mode is activated using the
power button and also upon SIM removal.

Domain: AOSP-Phone
Category: aosp improvement
Severity: medium
Origin: internal
Upstream-Candidate: yes

Change-Id: I88b267b12910fa37b9dff9e46ed639ee0b78fffc
Orig-Change-Id: I6cfa49afa92bbbe4337dc3a1192bc0d8c1b3ac7e
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/122682
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>