Phone: Disable SIM UIs in Airplane mode/SIM removal
BZ: 59396

If the airplane mode is activated using the power
button and if the current activity is one of the
SIM related activity, then the activity
should be disabled. Currently, it is not disabled.

Since the framework doesn't completely support
SIM access in airplane mode, there are few
issues seen. Same applies to if the SIM is removed
when the current activity is one of the SIM
related activity.

This patch disables/greys out the SIM related UIs
when the airplane mode is activated using the
power button and also upon SIM removal.

Change-Id: I6cfa49afa92bbbe4337dc3a1192bc0d8c1b3ac7e
Origin-Change-Id: Id6e7c16fd4fe10315bd6d06b8aee33d173fbb502
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Bin Yang <bin.y.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/76552
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>