[PORT FROM MAIN] Phone: Handling Network personalization PUK
BZ: 18880

Network personalisation is the feature were a SIM is restricted
to be used in a specific operator networks only. If an unsupported
SIM is used in the device, the user is asked to enter an unlock
code, so that the SIM can be used in the network. If the user enters
wrong PIN for a preprogrammed number of times, the device is PUK
locked and needs to be taken to the customer care. Android by default
supports entering the PIN, but doesn't inform the user if the device
is PUK locked.

This patch adds the handling of NETWORK PUK lock error in Phone App
and dismisses the PIN entry dialog.

Domain: AOSP-Phone
Category: aosp improvement
Severity: medium
Origin: internal
Upstream-Candidate: yes

Change-Id: I25438e6c1d8b252da8c575a42c450306153f7fa2
Orig-Change-Id: I6c09210f2450cfac7fc9962de02c297e0118cb99
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/122678
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>