[PORT FROM MAIN] Phone: Handling Network personalization PUK
BZ: 18880

Network personalisation is the feature were a SIM is restricted
to be used in a specific operator networks only. If an unsupported
SIM is used in the device, the user is asked to enter an unlock
code, so that the SIM can be used in the network. If the user enters
wrong PIN for a preprogrammed number of times, the device is PUK
locked and needs to be taken to the customer care. Android by default
supports entering the PIN, but doesn't inform the user if the device
is PUK locked.

This patch adds the handling of NETWORK PUK lock error in Phone App
and dismisses the PIN entry dialog.

Change-Id: I6c09210f2450cfac7fc9962de02c297e0118cb99
Orig-Change-Id: I99a1cbacb6f4cc84674e70f587c18010d4b9331c
Orig-Change-Id: Iac38c7a4b7d2199c189d21bd0cd97b85650edfa6
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/78490
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>