[PORT FROM MAIN] Phone: Change the headset plug/unplug detection
BZ: 40239

Previously, the PhoneApp detect the headset plug/unplug state
just by the value of intent state, it may cause error if the
headset plug is a kind of headset/headphone two in one plug.
As the definition from Google, we need to detect it according
to the value of both "state" and "microphone" element of the
intent.

Category: Device Enablement
Domain: Telephony-Modem-Intel
Origin: Internal
Upstream-Candidate: No,proprietary

Change-Id: I9abeb025b232c0a08640b3d0ec6cf18cae9e26ad
Orig-Change-Id: I095d7172f5f7c1915bdb62190e84bed86e50d267
Signed-off-by: Leo Yan <leo.yan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/124907
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>