Phone: Change the headset plug/unplug detection
BZ: 40239

Previously, the PhoneApp detect the headset plug/unplug state just
by the value of intent state, it may cause error if the headset plug
is a kind of headset/headphone two in one plug. As the definition
from Google, we need to detect it according to the value of both
"state" and "microphone" element of the intent.

Change-Id: I095d7172f5f7c1915bdb62190e84bed86e50d267
Orig-Change-Id: Ifcde6af139c80b33224799bab3eb200145e23799
Signed-off-by: Leo Yan <leo.yan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/85400
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>