PHONE: No ringtone screen when a call is coming with BTSCO activated
BZ: 58308

When a new call is incoming, the state of bt will be changed so that
a notificationis sent to notificationMgr to update the icon of bt state firstly.
Shortly, a new notificiation will be sent to notificationMgr to display the CallScreen.
The noficiationMgr will check the id of two notifications.
The id of two notifications are the same, so that the InCallScreen will not display.
Cancel the old notification before sending the new notification when there is a incoming call.

Change-Id: Ib0009def997043bed7387700fda6a7a4b0e8ea7d
Orig-Change-Id: I12a3562ea14b39c864c4977766c141aee2c1ab37
Signed-off-by: Bin Yang <bin.y.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/89772
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Diop, PiotrX <piotrx.diop@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>