[PORT FROM MAIN]PHONE: No ringtone screen when a call is coming with BTSCO activated
BZ: 58308

When a new call is incoming, the state of bt will be changed so that
a notificationis sent to notificationMgr to update the icon of bt state firstly.
Shortly, a new notificiation will be sent to notificationMgr to display the CallScreen.
The noficiationMgr will check the id of two notifications.
The id of two notifications are the same, so that the InCallScreen will not display.
Cancel the old notification before sending the new notification when there is a incoming call.

Feature: AOSP-Phone
Category: Aosp improvement
Severity: Medium
Origin: Internal
Upstream-Candidate: Yes

Change-Id: Ib7296c7dd8823e78cc65b4fa469b5d467263c96f
Orig-Change-Id: Ib0009def997043bed7387700fda6a7a4b0e8ea7d
Signed-off-by: Bin Yang <bin.y.yang@intel.com>