PHONE: Fix NullPointerException in voip call
BZ: 75259

In OnCreate function of InCallscreen.java, setInCallScreenInstance function
is called before the mInCallControlState is initialized.
Upon disconnect, audio and proximity sensor mode's will be updated.
While upding the speaker state, updateProximitySensorMode is called which
checks the value of mInCallScreen and use the mInCallControlState object.
If the updateProximitySensorMode function is called before mInCallControlState
is initialized that will result in the crash.
This patch fix this issue by moving mApp.setInCallScreenInstance() function to
the end of OnCreate().

Change-Id: Id7e9b9855f88d807323cdd8bf6f4a420d8850a8a
Orig-Change-Id: I0fcc0ef7816e1bd9321e7cfef344140692d17da3
Signed-off-by: Bin Yang <bin.y.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/88094
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>