[PORT FROM R42S] Phone: Improve placing emergency call in airplane mode
BZ: 94402

[DO NOT UPSTREAM]

EmergencyCallHelper issues dial request if the
service state is STATE_POWER_OFF. This results in
vendor ril receiving the dial request when the radio
is not available. Also a retry timer of 5seconds is
maintained which is not sufficient to bring up
the modem from powered off state.

In order to improve the emergency call success
rate following changes are done
- Increase the retry timer from 5 to 8seconds
- Allow placing emergency call if the service
state is not powered off and in service or
emergency call or if the sim state is other than
unknown.

Further improvements needed.

Change-Id: Ie6563359b570896b907506a6d2a87228d2b7076d
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/102875
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>