Phone: improve the 'additional settings' feature to avoid wrong status
BZ: 71111

In the call settings, if we go to the 'Additional settings' entry we have
the possibility during a short time to press one of the options before the
displaying of the updating pop-up. If we do this, we will have a wrong
status for the setting values.

This patch updates the code to have the pop-up displayed immediatly as
it's done for call forward options.

Change-Id: Ia1e087ab343036ea18e4ac5abeacdbdfd8c0eec3
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/110211
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Daragnes, Julie <julie.daragnes@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>