[PORT FROM R42S] Phone: dismiss the depersonalization panel if the SIM card is removed
BZ: 111860

The depersonalization panel to enter the network PIN code is
still visible when we removed the SIM card. This can lead to
a crash of the com.android.phone process if the user tries
to enter a PIN code at this time.

This patch fixes this issue by displaying a message if the
SIM card is removed before to dismissing the panel.

Change-Id: Icef7c2c8853528a1918afadf46f7508071b7149d
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/111838
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>