[PORT FROM MAIN] Phone: dismiss the depersonalization panel if the SIM card is removed
BZ: 111860

The depersonalization panel to enter the network PIN code is
still visible when we removed the SIM card. This can lead to
a crash of the com.android.phone process if the user tries
to enter a PIN code at this time.

This patch fixes this issue by displaying a message if the
SIM card is removed before to dismissing the panel.

Category: AOSP Improvement
Domain: Telephony-Phone
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I397e126390497bfd9b4232346c6406db03918a00
Orig-Change-Id: Icef7c2c8853528a1918afadf46f7508071b7149d
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/125146
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>