Phone: Fix radio off request not sent to framework
BZ: 123894

[DO NOT UPSTREAM]

If the platform shutdown is initiated just after
the radio is powered on, modem is not powered off
properly.

When the radio power on is ongoing, service state
still remains in STATE_POWER_OFF leading to platform
shutdown even with the modem still powered on.

Allow the radio power request to be sent always
to the telephony framework irrespective of the radio
and service state.

Change-Id: I065af4c986867b335f708cd1e91199df86168a0e
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/119808
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>