[PORT FROM R42BS] Disable speakerphone when connecting BT device in call
BZ: 125510

When connecting a BT device in call, the BT stack sends to
AudioService a setForceUse BT_SCO message. If the phoneApp
had previously sent a message to force the Speaker in call,
the AudioService will choose arbitrarily the device that the
communication has to be forced on. This can result in keeping
the speaker as the output device.

Since the phoneApp is fully aware of the BT state change, the
purpose of this fix is to prioritize (just as it was done until
now for wired devices) BT usage over speaker. This way the
phoneApp will explicitly inform the AudioService to unset the
previously asked forced usage for speaker.

Category: Device Enablement
Domain: CWS.BT-Audio
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I2f9da788c9276b5b9ea8b14318cb1da061eaae3a
Signed-off-by: Piotr Diop <piotrx.diop@intel.com>
Reviewed-on: http://android.intel.com:8080/127346
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>