Disable speakerphone when connecting BT device in call
BZ: 125510

When connecting a BT device in call, the BT stack sends to
AudioService a setForceUse BT_SCO message. If the phoneApp
had previously sent a message to force the Speaker in call,
the AudioService will choose arbitrarily the device that the
communication has to be forced on. This can result in keeping
the speaker as the output device.

Since the phoneApp is fully aware of the BT state change, the
purpose of this fix is to prioritize (just as it was done until
now for wired devices) BT usage over speaker. This way the
phoneApp will explicitly inform the AudioService to unset the
previously asked forced usage for speaker.

Category: Device Enablement
Domain: CWS.BT-Audio
Origin: Internal
Upstream-Candidate: Yes

Change-Id: Iab69cfd3152636d91c605896982bca79e4b0bbb7
Signed-off-by: Piotr Diop <piotrx.diop@intel.com>
Reviewed-on: http://android.intel.com:8080/121901
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yin, Fengwei <fengwei.yin@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>