[PORT FROM MAIN] Telephony: Call Forwarding items are grayed when wrong numbers are used
BZ: 52109

In call forwarding settings (always, busy, unanswered, unreachable), when user
enter an empty or too long phone number, a pop-up with an error message
appears and button is grayed.
If user wants to modify the settings again, it is necessary to go in sleep
mode or go back and re-enter in the menu.
With this patch, pop-up is still available and button is never grayed.
When user enter a wrong phone number, button is still available

Category: AOSP Improvement
Domain: Telephony-Phone
Severity: Medium
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I433b1120a08a62f3e0a9f5060580fd6d26b2d0b9
Orig-Change-Id: Iccc29bd13aa05370f8680caa1f80c58f6dc6c812
Signed-off-by: Goubert, Tony <tonyx.goubert@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/124602
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>